[sival,kmac] Add kmac_endianess test
This commit adds a chip level test to test the endianess feature
of KMAC. This includes the message as well as the state register.

The following configurations are covered:
- Little endian message, little endian digest.
- Little endian message, big endian digest.
- Big endian message, little endian digest.
- Big endian message, big endian digest.

Currently, only the KMAC mode is covered. However, as the big/
little endian feature should be independent of the underlying
mode, this should be sufficient.

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>