[manuf] Calculate OTP measurements in ft_personalize
This calculates the measurement of the OTP partitions in
`ft_personalize` and pass them to `dice_uds_tbs_cert_build()`.
This change is necessary because the CreatorSwCfg and OwnerSwCfg might
not be fully provisioned to the OTP when building the UDS certification.
By calculating the final measurements in `ft_personalize`, we can get
the values of unprovisioned field from the OTP image to ensure the
correct values are used in the certificate generation process.

Signed-off-by: Anthony Chen <antchen@google.com>