[opentitantool] Add vcd outfile option to gpio monitor start and read
Augment the gpio monitoring start and read command with the ability to
also write a vcd file. The start and read commands do not block or hold
the opentitantool interface, so they can be used in conjunction with
other opentitantool commands. For example, the gpio monitoring start and
read commands can be used with the tpm i2c commands to record all of the
i2c traffic from hyperdebug and the DUT. The existing gpio monitoring
vcd command holds the hyperdebug resource until it is finished, so it is
currently not possible to sent tpm i2c commands to the DUT using
opentitan tool while the gpio monitoring vcd is running.

Change-Id: Ideca33ad23afefe1dad718c97ed3a1dea992897b
Signed-off-by: Jett Rink <jettrink@google.com>