[manuf] Bind perso firmware with INITIAL MANUF_STATE
This binds the personalization firmware with the `PERSO_INITIAL`
manufacturing state. This ensures the personalization firmware's
signature is invalidated once the device is fully personalized and
prevents the personalization firmware from being executed again.

A new manifest, `manifest_perso` is created with the `PERSO_INITIAL`
creator manufacturing state to facilitate this binding.

Signed-off-by: Anthony Chen <antchen@google.com>