[dv, dvsim] : Override NC node's TMPDIR with custom value
Bazel creates sandboxes and mounts the TMPDIR in this sandbox. When
bazel is launched as part of the build process for a test that has been
forked off to a NC node, the sandbox created in the process has access
issues to the tmp/tmpfs mount provided by the NC node using the default
TMPDIR environment variable. This workaround overrides the TMPDIR to a
location under the project workspace which is on the shared NFS

Signed-off-by: Venkat Krishnan <venkateswarank.opentitan@gmail.com>