[pinmux] Remove users of hw/ip/pinmux/data/pinmux.hjson
This file should not be used: the actual hsjon file is created
by topgen based on a template but this file is used during the
topgen process as a "default value file". This is a hack due
to the fact that pinmux is not a proper IP template.
Using this file could lead to using outdated data.
There were only two references to
it that this commit points to the correct location. This commit
also removes two bazel targets for headers based on this file.

Signed-off-by: Amaury Pouly <amaury.pouly@lowrisc.org>