[dv, aes] : Relax strict wall timeout for aes fi tests
Increasing timeout from 1 min to 5 mins for fi tests as 1 min run times
are not practical when running on compute farm with heavy workloads from
multiple projects operating simultaneously. The wall time includes the
time spent in a queue waiting to land on a node in the farm and possibly
the time to acquire a simulator license. Then depending on workloads on
the servers and the NFS disk, the job may sometimes run slower
(clock cycles per sec) than usual. Increasing the timeout value to avoid
false failures.

Signed-off-by: Venkat Krishnan <venkateswarank.opentitan@gmail.com>