[bazel,otlib] enable otlib to be built as external dep
This change enables this repo (`lowrisc_opentitan`) to be brought in as
an external Bazel dependency in another Bazel project for the purpose of
building opentitanlib as an external dependency in another Bazel
project. This will be used by the `opentitan_provisioning` repo to build
a provisioning testbed based on FPGA provisioning flows.

Signed-off-by: Tim Trippel <ttrippel@google.com>