[rom_ctrl, dv] Fixed tlul_adapter_reg uncovered conditional coverage
The nightly coverage dashboards were reporting two uncovered conditional
coverage in tlul_adapter_reg.sv at L208. There is a conditional assignment
on L208 for err_internal, which is getting assigned if any of the bits
(addr_align_err, malformed_meta_err, tl_err, instr_error and intg_error)
is set. The case 0 | 0 | 0 | 1 | 0 was uncovered as the assignment for
tl_err, instr_error and malformed_meta_err are depending over common
factors.
It turns out that instr_error could exist on its own if we have an opcode
for A channel as GET, instr_type as MuBiTrue and fetch for register is
disable.
This fix added a mechanism inside cip_base_vseq__tl_errors.svh to mimic
the above behaviour and then let the scoreboard knows that it should
expect an error and raise an invalid access.

Signed-off-by: Kinza Qamar <kqzaman@lowrisc.org>