[manuf,otp] provision secure boot keys during individualization
This updates the FT individualization provisioning stage to provision
the first stage secure boot keys into the ROT_CREATOR_AUTH* partitions.

Doing so required updating the OTP C file constants generation to allow
generating a C file that contains the secure boot key constants that are
to be provisioned into OTP during individualization.

Additionally, this fixes a bug in the generation of the perso firmware
that was always using the sival SKU OTP SW_CFG constants for the fields
that cannot be provisioned during individualization. Now, a separate
personalization binary is generated for each SKU.

This fixes #21554.

Signed-off-by: Tim Trippel <ttrippel@google.com>