[pwm,dv] Cap number of sequences in pwm_stress_all_vseq
Fixing the number of sequences that is run to 5 (instead of uniformly
choosing from [5, 8]) will slightly reduce the average number of
iterations (from 6.5 to 5). The plus side is that it will avoid very
long runs (which occasionally time out because they take more than
10s). We'll still get a reasonable number of "back-to-back"
sequences (since 5 is much more than 1!).

The total amount run will drop slightly. To avoid that, we could up
the reseed count for this test to 6.5/5*50 = 65, but I'm reasonably
convinced that the number for this vseq come from a default, so I
don't believe there's any need to worry about it.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>