[pwm,dv] Rewrite pwm_scoreboard using out-of-block declarations
No functional change, but this allows us to attach proper
documentation comments to the functions/tasks and hopefully make the
class a bit easier to understand at a glance.

This commit also gets rid of get_reg_index, which was replaced by
get_multireg_idx from pwm_env_pkg with 92c5733 in 2021.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>