[pwm,dv] Rewrite pwm_env_cfg using out-of-block declarations
Since we were touching the code anyway, I've also changed the way that
we randomise the core frequency (divided down from the bus frequency).
The code looked like the scaling was configurable, but it is always
enabled. It also re-randomised on every call of what looks like a
getter (after the rest of the class had been randomised).

This commit changes things so we randomise the scaling ratio (picking
a value between 1/4 and 1) and then apply it with the call to
get_clk_core_freq. This two-stage process is needed because we
randomise the cfg object before connecting up the vif that determines
the bus frequency.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>

[pwm,dv] Remove unused en_random argument from get_clk_core_freq

This is never supplied at the call sites (and I don't think it ever
has been since it was defined in 2021!).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>