[pwm,dv] Remove set_reg_en from pwm_base_vseq
This task was supposed to work by setting the REGWEN to update the
state. This is a bit odd, since you can only ever clear REGWEN. What's
worse, the DV code was written when the register was rw1c (so it
cleared the register by writing 1).

The register become rw0c in 2021 with commit 25e147d98eb, so this code
has done nothing since then.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>