[sdc] Add clarifying comments to SDC file, move SPI HOST 1 PT waivers
Some of the constraints in the SDC file were not fully clear and this
commit adds comments to provide relevant background information. In
addition, the PrimeTime set_case_analysis waivers for SPI HOST 1 are
moved to the section applied in PrimeTime only.

This is a follow-up of lowRISC/OpenTitan#24221.

This is a cherry pick of commit c03805b580ca3b8acba5572b2f09638807060950
to branch earlgrey_1.0.0.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>