[sdc] Add clarifying comments to SDC file, move SPI HOST 1 PT waivers
Some of the constraints in the SDC file were not fully clear and this
commit adds comments to provide relevant background information. In
addition, the PrimeTime set_case_analysis waivers for SPI HOST 1 are
moved to the section applied in PrimeTime only.

This is a follow-up of lowRISC/OpenTitan#24221.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>
(cherry picked from commit 567f3b35fddbc07b1952ed27f4ed0f269caeafd2)