[rom_ext] encapsulate dice logics for moving to silicon_creator lib
This refactors the DICE-related logics in rom_ext.
Global variables are encapsulated in `dice_cert_t` and related code are encapsulated in `dice_chain_*` functions.

It also includes more boundary checks on the cert buffer in RAM to avoid overflow.

Also, tlv object is parsed at where it will be consumed to handle parsing error correctly,
and stale value will be cleared if parsing failed.

Signed-off-by: Yi-Hsuan Deng <yhdeng@google.com>