[ci,check-generated] Handle empty text in CMDGEN
Change ci/scripts/check-generated.sh to run both top and cmdgen
when checking for the topgen generated artifacts. This is to
support template md files having no text between the BEGIN and
END markers so they can handle multi-top.

Add a separate "cmdgen" target in the hw/Makefile and describe
that in the README.md for contributing to documentation.

Fixes #24991

Signed-off-by: Guillermo Maturana <maturana@opentitan.org>