[manuf,rom_ext] move UDS cert to separate info page
The DICE UDS certificate is endorsed offline at manufacturing time, and
injected into a flash info page on the chip. It must remain stable for
the lifetime of a chip.

Previously, the CDI_0 and CDI_1 DICE certificates were also written to
the same flash info page. However, the ROM_EXT can update these
certificates in the event that the ROM_EXT or Owner FW is ever updated
(as this triggers a CDI0/1 key pair change). To update these
certificates, the ROM_EXT buffered all certificates in RAM, erased the
flash info page, and re-wrote it with the udpated certificates.

If there is a power outage during the certificate update, after the
flash page was erased but before the certificates had been written back,
there could be a moment when the UDS certificate was lost.

To prevent this from happening, the UDS certificate was moved to its
own flash info page, that is never erased for the lifetime of the chip.
Additionally, this updatest the ROM_EXT to never write a dummy UDS cert
to flash in the event a chip is not provisioned, or we are running on a
sim/FPGA platform, and instead, just write a warning message to the
console.

Signed-off-by: Tim Trippel <ttrippel@google.com>