[kmac, rtl/dv/sw] Remove unused error code
The `ErrShadowRegUpdate` code is never used in the RTL. Instead, according
to the documentation, on a shadow register update error, a recoverable alert
shall be triggered - this recoverable alert then is signaled in the STATUS
register and not in the error code register.

Hence, this commit removes this error code as well as the corresponding
SW code (`kDifErrorShadowRegisterUpdate`).

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>