[manuf,rom_ext] Update the parameters of perso_tlv_cert_obj_build
When building a perso_blob, the object type was decided by whether that
object needs endorsement or not. It is not feasible when not only one
certificate format needs to be supported by perso_fw and ROM_EXT.
To remediate that it requires the specify the object format by the
caller which needs to add one more parameter.

Signed-off-by: Tommy Chiu <tommychiu@google.com>