[dv,alert_esc_agent] Tighten up the logic in alert_receiver_driver
The existing structure of this agent was pretty confusing because it
looked like an alert response item (as driven by this class) would
hang until an alert request came in.

In practice, this doesn't happen because we don't actually generate
the response item until the request has come in, but I didn't find it
particularly obvious.

With this rewrite, we fail in an obvious way if we get given a
response item without a request.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>