[dv, spi_dev] model+driver+seq_item fix for read pipeline
The spi-host driver wasn't feeding the read_pipeline delay when
driving read commands. The driver now waits for read_pipeline delay when
reading from the read buffer.

The scoreboard watermark interrupt and last_read_address needed to be
updated accordingly for read commands targetting the buffer since the
transaction can be now up to 2-clock cycles longer if a command is
targetting the read buffer

Signed-off-by: Antonio Martinez Zambrana <amz@lowrisc.org>