[dv] Exclude some secded generation modules from coverage
This is slightly extending commit 803eed8, which excluded two secded
decoders. For similar reasons, we might want to exclude the
corresponding encoders. Again, these are verified in FPV and getting
complete conditional coverage is hard (or maybe impossible) to hit. No
need to spend time on doing so: the FPV should catch any bug in these
prims.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>