[pwm,dv] Manually exclude some coverage in prim_reg_cdc instances
This expression coverage is expecting to see a src_update signal equal
to 1. That signal can only be nonzero if the DstWrReq parameter is 1,
but it is wired to zero.

A working UNR flow would waive this but we don't have that set up
properly in a way that I can run it right now, so I'm doing the manual
exclusion instead.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>