[manuf] verify manuf_state_creator consistency between ROM_EXT and OTP
This adds the check to ensure the `manuf_state_creator` field in the
ROM_EXT manifest matches the expected value that will be provisioned
into the OTP. The ROM_EXT manifest is located in slot b and is flashed
during the second bootstrap in the personalization flow. The
verification is done before the `CREATOR_MANUF_STATE` and
`ROM_BOOTSTRAP_DIS` OTP fields are provisioned, allowing for correction
and re-running the personalization flow.

Signed-off-by: Anthony Chen <antchen@google.com>
(cherry picked from commit af9c2695fa6680682a868281d0cd449c3972541d)