[rom_ext,dice] Fix incorrect dice page id.
There's one rebase conflict resolved incorrectly in #24980.
It tries to load the CDI certs from the Factory page which is only for UDS cert.
So the certificate is always not found, making the rom_ext refreshing the cert
chain every boot. This fixes said bug and ensures DICE CDI_* certs are
only updated when they are detected to be invalid.

An new e2e test is added to ensure we skip unnecessary update correctly.

Signed-off-by: Yi-Hsuan Deng <yhdeng@google.com>