[pwm,dv] Exclude toggle coverage for some constant fields of tl_o
For the first item (d_param and d_sink), the fields are genuinely
constant.

For the second item, the rsp_intg signal is calculated as a hash of a
small value. Local attempts to change it, watching a wave viewer,
didn't actually see the top bit change. I'm not sure if it's actually
impossible to hit, but waiving the coverage seems reasonable in this
case.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>