[tl,dv] Avoid silly hole in tl_a_chan_cov_cg
The cp_mask field expects to see a value of '1 (i.e. all byte lanes
are enabled). Bizarrely, a full-width read doesn't match the
coverpoint bin for some reason, yielding a hole in the cross at the
bottom of the covergroup.

Calculate the "replicated ones" explicitly, which seems to solve the
problem.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>