[rv_core_ibex, top_earlgrey] Enable SecureIbex for CW340
As the `SecureIbex` configuration is the one we use for the
Earl Grey ASIC, we also should enable it for the FPGA boards.

Due to resource constraints, this is only possible for the CW340
but not the CW310.

Resource utilization stays still will below 70%.

Closes lowRISC/opentitan#25137

Signed-off-by: Pascal Nasahl <nasahlpa@lowrisc.org>
(cherry picked from commit ce82d8008a6176ddb455accdac486fba79741620)