[opentitantool] HyperDebug driver robustness
When the driver reads a command response from the HyperDebug text
console, it no longer relies on getting timeout to recognize when the
output is complete.  (Instead if it sees the prompt "> " as the final
two characters in any chunk received, it is all.)

With that in mind, there is no reason to have the timeout as low as
100ms.  In particular the `reinit` command executed as part of
"transport init" takes a little while to calibrate ADC and reset various
circuits in HyperDebug, and could get more responsiblities in the
future.  And on top of that, the USB bus introduces sometimes
unpredictable delays.

This CL increase the general delay for HyperDebug console response to 3
seconds.  And also propagates any error messages from `reinit` in a
slightly better way.

Change-Id: I3ecb17a3ad32f923b240d0dcce76e6d40938e4c8
Signed-off-by: Jes B. Klinke <jbk@chromium.org>
(cherry picked from commit c5d704716c0e4a4a6e1d5cb8d5311b8413b8b78e)