[ci] Add checkout step to verify-fpga job
Without this, Azure automatically checks out the repo at the time this
job runs, not necessarily at the same commit as the previous FPGA jobs.

This can lead to a test not existing when the FPGA tests ran, but now
existing when this verify job runs. The verify job then complains that
you didn't run that test.

Signed-off-by: James Wainwright <james.wainwright@lowrisc.org>