[sival] Remove unused OTP targets from sival SKU.
The provisioning firmware contains hardcoded settings for the HWCFG1
partition, and uses data generated at manufacturing time to calculate
the DEVICE_ID. As a result, most of the otp image targets and hardware
partition settings in the sival SKU, are not derived from the OTP build
targets.

This change removes unused tardgets to simplify the build / provisioning
maintenance. pre-silicon and post-silicon simulation and emulation
environments should use the emulation SKU instead.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>