[kmac,sival] Enable KMAC error conditions test to partially run on CW310
This commit enables the KMAC error conditions test in the CW310 FPGA
execution environment, conditionally disabling the first three tests
when running in this environment - the CW310 lacks the masking features
for these tests to run, but there is still value in running the
remaining 7 tests in a CW310 environment. CI prioritisation now also
prioritises CW340 over CW310, meaning that the full set of tests will
still be running in CI as usual - this just helps provide more
thorough testing.

Signed-off-by: Alex Jones <alex.jones@lowrisc.org>