[manuf] refactor CA args to personalization flow
This creates a CaConfig struct with CA args, and moves it into the
cert_lib to better consolidate the configuration settings for endorsing
certificates during personalization. Additionally, this refactors how the
CA arguments are passed to the provisioning flows. This is in preparation
for supporting multiple CAs for endorsing different certificate chains.

This partially addresses #24955.

Signed-off-by: Tim Trippel <ttrippel@google.com>