[pwm,dv] Teach pwm_monitor to handle resets and cfg.active properly
The change in behaviour should only ever matter if we see things that
look like transactions when we're under reset or the config doesn't
think it is active. Practically speaking, this isn't going to happen,
but the change should hopefully make the code easier to reason about.

The change also adds some documentation comments to explain how the
monitor is expected to work and restructures the logic to be a bit
easier to follow.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>