[opentitantool] HyperDebug support for advanced TPM primitives
Add support in the HyperDebug transport driver for `TpmPoll` and
`GscReady`.  It will reject the opertations if not supported by
HyperDebug firmware.  Though that is only going to be the case if the
user has failed to issue `transport update-firmware`, or is using this
backend driver with some other partially-compatible debugger devices.
(Also known as "flavors" of HyperDebug.)

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I154229794bcb88cc7845329c03621eec5c4da4a5