[opentitantool] Enhance SPI and I2C traits for TPM
Add a `set_pins()` method to the I2C trait.  This allows instructing the
debugger to set up its pinmux, if it has such support, in the same way
as the existing method on the SPI trait.

Also, add a new pin to both SPI and I2C called `gsc_ready`.  This one
will not be used by normal I2C or SPI communication, but the debugger
can be instructed to wait for a falling edge on that pin, before
continuing the transaction, typically the sequence given to
`run_transcation()` would look somewhat like `[Write(...), GscReady,
Read(...)]` which would have the effect that the debugger would send a
number of bytes, and then wait for falling edge before resuming clocking
to retrieve the response.  This is useful when communicating with Google
Security Chips.

Finally, the SPI trait is enhanced with a `TpmPoll` instruction, which
is to be used like `[Write(...), TpmPoll, Read/Write(...)]` when
communicating with standards-compliant TPM devices.  It has the effect
that the debugger will first send the request (must be 4 bytes), then it
will repeatedly poll by reading a single byte at a time, until the
device responds with 0x01, before the debugger reads or writes the main
data part.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>

Change-Id: Ied56bac014a6d5bfe501a6bd3f34c7b21b69265d