[alert,dv] Fix logic in alert_receiver_driver
This was "tightened up" in commit b1095ff, which actually broke
things! The problem is that I was imagining the "continue" would
continue in the surrounding "forever" loop. But it doesn't! It just
skips an iteration of the "repeat" loop inside.

Make the logic a bit more explicit and correctly handle the case where
we go into reset with the clock paused.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>