[opentitantool] Optimization, avoid excessive Ctrl-C
HyperDebug has a text-based console, which opentitantool/lib uses for
sending all sorts of commands.  For the case a previous invocation has
left a partially typed command in HyperDebug's buffers, opentitantool
always precedes any command with Ctrl-C to clear the buffer.

This is slightly inefficient, when opentitantool is sending many
commands in succession, and has locked the character device preventing
any other Linux process from injecting any input.

This CL changes the code to only emit Ctrl-C before the first time a
command is sent while the lock is held.

Change-Id: Id209f5ef5125ec2bb3f0489c33b67a291207cf6b
Signed-off-by: Jes B. Klinke <jbk@chromium.org>