Create usbdev_toggle_restore_test
Test for all targets that exercises the Data Toggle save/restore
functionality that was introduced to support resuming from
Deep Sleep. This test will not pass on ES since the necessary
logic and registers are not present.

It requires a physical connection to a USB host controller,
or the presence of the DPI model in Verilator/ASIC simulation,
but there is no required communication with the host/DPI.

Only the presence of VBUS and taking the link out of reset is
required, in order to be able to set the Data Toggle bits
from software.

Signed-off-by: Adrian Lees <a.lees@lowrisc.org>